API

SlidingDFTs.dataoffsets — Method

dataoffsets(method)

Return an integer or a vector of integers with the offsets of data samples that are needed by the given method to compute a sliding DFT.

If the code of SlidingDFTs.updatepdf! that dispatches on the type of method uses the function SlidingDFTs.previousdata, this function must return the integers that are used as the third argument (offset) of that function. If that function is not needed (no past samples are used), this one may return nothing to reduce memory allocations.

SlidingDFTs.dftback — Method

dftback(method)

Return an integer or a vector of positive integers with the indices of the previous iterations that are needed by the given method to compute a sliding DFT.

If the code of SlidingDFTs.updatepdf! for the type of method uses the function SlidingDFTs.previousdft, this function must return the integers that are used as the third argument (back) of that function.

If that function is not needed, this one may return nothing to reduce memory allocations.

SlidingDFTs.iterationcount — Method

iterationcount(state)

Return the number of iterations done for the sliding DFT represented by state.

If the most recent iteration computed the DFT of the fragment of x corresponding to the range i : i+n, then this function returns the number i.

SlidingDFTs.nextdata — Method

nextdata(state)

Return the next value after the fragment of the data series that was used in the most recent iteration of the sliding DFT represented by state.

If the most recent iteration computed the DFT of the fragment of a data series corresponding to the range i : i+n, then this function returns its i+n+1-th value.

There is no defined behavior if such value does not exist (i.e. if the end of a finite data series was reached).

SlidingDFTs.previousdata — Function

previousdata(state[, offset=0])

Return the first value of the fragment of the data series that was used in the most recent iteration of the sliding DFT represented by state, or at offset positions after the beginning of that fragment.

If the most recent iteration computed the DFT of the fragment of a data series corresponding to the range i : i+n, then this function returns its i+offset-th value.

SlidingDFTs.previousdft — Function

previousdft(state[, back=0])

Return the DFT computed in the most recent iteration of the sliding DFT represented by state, or in a number of previous iterations equal to back.

If the most recent iteration computed the DFT of the fragment of a data series corresponding to the range i : i+n, then this function returns its DFT for the range i-back : i+n-back.

SlidingDFTs.stateful_sdft — Function

stateful_sdft(method, x[, safe=true])

Return an iterator to produce a sliding DFT of x, using the given method, considering that x is a stateful iterator.

See sdft(@ref) for more details.

SlidingDFTs.updatedft! — Function

updatedft!(dft, x, method, state)

Update the values of a sliding Discrete Fourier Transform (DFT) of a data series, according to the algorithm of the provided method, for a given state of the sliding DFT.

dft is a mutable collection of complex values with length equal to windowlength(method), containing the value returned by the last iteration of the sliding DFT.

x is the data series for which the sliding DFT is computed, at least as long as windowlength(method).

method is the object that defines the method to compute the sliding DFT.

state is an object generated automatically at each iteration of an object created by sdft(method, x) or stateful_sdft(method, x), containing the information that is needed to compute the new values of the sliding DFT, together with x. That information can be extracted from state with the following functions:

SlidingDFTs.nextdata(@ref) to get the next value of the data series.

SlidingDFTs.previousdata(@ref) to get a previous value of the data series.

SlidingDFTs.previousdft(@ref) to get the DFTs of previous iterations.

SlidingDFTs.windowlength — Function

SlidingDFTs.windowlength(method)

Return the length of the window used by method.